Migration now working. Migration is activated when a CPU deadline is missed.

+++ DistributedSystemSimulator/CodeBase.cpp
@@ -6,7 +6,7 @@
 // CodeBase Class implementation below
 //////////////////////////////////////////////////////////////////////////
 
-bool CCodeBase::AddCommand(bool i_memoryAccess, uint64_t i_cmdCount /*= 1*/, uint64_t i_startAddr /*= 0*/)
+bool CCodeBase::AddCommand(bool i_memoryAccess, uint32_t i_cmdCount /*= 1*/, uint32_t i_startAddr /*= 0*/)
 {
 	CCommand command(i_memoryAccess, i_cmdCount, i_startAddr);
 	m_commandVector.push_back(command);
@@ -23,9 +23,9 @@ uint16_t CCodeBase::GetCommandCount()
 	return count;
 }
 
-uint64_t CCodeBase::GetMaxAddr()
+uint32_t CCodeBase::GetMaxAddr()
 {
-	uint64_t maxAddr = 0x0;
+	uint32_t maxAddr = 0x0;
 	for (std::vector<CCommand>::iterator it = m_commandVector.begin(); it != m_commandVector.end(); ++it) {
 		if (maxAddr < it->GetMaxAddr())
 			maxAddr = it->GetMaxAddr();
@@ -44,7 +44,7 @@ bool CCodeBase::IsCommandMemAccess(uint16_t i_commandIndex)
 	return false;
 }
 
-uint64_t CCodeBase::GetAddressIndex(uint64_t i_commandIndex)
+uint32_t CCodeBase::GetAddressIndex(uint32_t i_commandIndex)
 {
 	int signedCommandIndex = i_commandIndex;
 	int addressIndex = 0;
