+++ DistributedSystemSimulator/Command.h
@@ -4,14 +4,14 @@ class CCommand
 {
 public:
 	CCommand(){};
-	CCommand(bool i_memoryAccess, uint64_t i_cmdCount = 1, uint64_t i_startAddr = 0);
+	CCommand(bool i_memoryAccess, uint32_t i_cmdCount = 1, uint32_t i_startAddr = 0);
 	~CCommand(){};
 	bool IsMemoryAccess();
-	uint64_t GetStartAddr();
-	uint64_t GetCmdCount();
-	uint64_t GetMaxAddr();
+	uint32_t GetStartAddr();
+	uint32_t GetCmdCount();
+	uint32_t GetMaxAddr();
 private:
 	bool m_memoryAccess;
-	uint64_t m_startAddr;
-	uint64_t m_cmdCount;
+	uint32_t m_startAddr;
+	uint32_t m_cmdCount;
 };
